requirements.txt
celery==5.2.1
django-bootstrap4==21.1
django-ckeditor==6.2.0
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-import-export==2.6.1
django-meta==2.0.0
django-mptt==0.13.4
django-registration==3.2
django-solo==2.0.0
django-tables2==2.4.1
django-webpack-loader==1.4.1
Django==3.2.9
djangorestframework-api-key==2.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
feedparser==6.0.8
loguru==0.5.3
psycopg2-binary==2.9.2
pymorphy2==0.9.1
pytils==0.3
redis==4.0.2
requests==2.26.0
sentry-sdk==1.5.0
